Small data-driven test of QuickSort
Added a value-parameterized test fixture to test QuickSort on a bunch
of example arrays. Using static member variables to store the example
arrays and SetUpTestCase, TearDownTestCase to construct/delete them,
and passing them directly to the test, thus violating the rule that a
test should not modify static members of its test fixture. But this is
okay because there is only one test in the test case/fixture (and
different instances of it access different indices of the static array
of arrays).