New version of SortTestWithExamples: no "static" problem
Made a new version of SortTestWithExamples that gives each test a fresh
copy of the static example array. That way there is no longer the
restriction of having only one test using the class as the test
fixture. The old test fixture is also kept, renamed as
SortTestWithStaticExamples. Also, used vectors to make things a bit
cleaner.