SortTestWithPatterns: to test using patterns, e.g. ascending, descending, constant, of increasing lengths
Value-parameterized test fixture that takes in a struct parameter:
string name, int size. All-constant arrays is BAD for even the middle
pivot of QuickSort. Left/Right pivots for ascending or descending
arrays are obviously bad, but note there are nuances (e.g. left pivot
for ascending is not that bad). Also replaced NULL with nullptr.