Improvement to PartitionLomutoCLRS to avoid worst case on array with all-equal elements
All QuickSort variants so far, including LomutoCLRS, suffered
worst-case performance on array of all-equal elements. Made my own fix
for it by alternating on where the equal elements go… to the left in
even iterations and to the right in odd iterations. Implemented this by
unrolling the for loop of PartitionLomutoCLRS. Also extended the
SortTestWithExamples test case by adding the Lomuto CLRS variants to it.