QuickSort using Hoare Partition! The original, much efficient, harder to understand, but beautiful method of partitioning.
Hoare Partition is tricky to understand. But I think I got it. Even
able to write correct, meaningful, compact invariants. Hoare Partition
in two minor variations: one that returns one partition index, and one
that returns two partition indices. Also tried the Median of Three
method of picking pivot. Again, two minor variations of that: one where
the index of the median is used as the pivot, and one where the three
elements are re-arranged (i.e. sorted) to bring the median at the
middle index (which is then used as the pivot).