Insertion, Selection, Bubble sorts
Added InsertionSort, SelectionSort, and four variants of BubbleSort.
Also added their tests. The implementations closely match those in CLRS
book. Except for my own optimization to BubbleSort — keep track of the
last swap location in a pass, and jump the outer loop index to that
location for the next pass. This improvement gives O(n) performance on
already-sorted input, which is like InsertionSort.