Added QuickSort described in the K&R book on C
The K&R book on C uses the example of quicksort when introducing
recursive functions. It uses what it calls “not the fastest, but one of
the simplest” method of quicksort. It is in fact the Lomuto partition
scheme. But slightly different from that in the CLRS book. This one
swaps the pivot to the left and partitions left+1..right.