Binary Tree traversals without recursion
Pre-order using stack is easy. In-order and post-order are tricky.
Wikipedia (https://en.wikipedia.org/wiki/Tree_traversal) lists nifty
implementations using just a stack and one/two extra pointers. But my
initial thought was to track state - 0, 1, 2 - for a node that is on
the stack. Now, a nice progression of in-order implementations that
transform step-by-step from the initial stack+state solution to the
nifty one.